+++ hutool-bloomFilter/src/main/java/cn/hutool/bloomfilter/bitMap/IntMap.java
@@ -7,7 +7,6 @@ package cn.hutool.bloomfilter.bitMap;
  *
  */
 public class IntMap implements BitMap {
-	private static final int MAX = Integer.MAX_VALUE;
 
 	private int[] ints = null;
 
@@ -48,7 +47,7 @@ public class IntMap implements BitMap {
 	public void remove(long i) {
 		int r = (int) (i / BitMap.MACHINE32);
 		int c = (int) (i % BitMap.MACHINE32);
-		ints[r] = (int) (ints[r] & (((1 << (c + 1)) - 1) ^ MAX));
+		ints[r] &= ~(1 << c);
 	}
 
 }