+++ hutool-core/src/main/java/cn/hutool/core/convert/Convert.java
@@ -609,7 +609,40 @@ public class Convert {
 	public static <T> T convert(Type type, Object value, T defaultValue) throws ConvertException {
 		return ConverterRegistry.getInstance().convert(type, value, defaultValue);
 	}
-
+	
+	/**
+	 * 转换值为指定类型，不抛异常转换<br>
+	 * 当转换失败时返回{@code null}
+	 * 
+	 * @param <T> 目标类型
+	 * @param type 目标类型
+	 * @param value 值
+	 * @return 转换后的值，转换失败返回null
+	 * @since 4.5.10
+	 */
+	public static <T> T convertQuietly(Type type, Object value) {
+		return convertQuietly(type, value, null);
+	}
+	
+	/**
+	 * 转换值为指定类型，不抛异常转换<br>
+	 * 当转换失败时返回默认值
+	 * 
+	 * @param <T> 目标类型
+	 * @param type 目标类型
+	 * @param value 值
+	 * @param defaultValue 默认值
+	 * @return 转换后的值
+	 * @since 4.5.10
+	 */
+	public static <T> T convertQuietly(Type type, Object value, T defaultValue) {
+		try {
+			return convert(type, value, defaultValue);
+		} catch (Exception e) {
+			return defaultValue;
+		}
+	}
+	
 	// ----------------------------------------------------------------------- 全角半角转换
 	/**
 	 * 半角转全角
@@ -664,6 +697,9 @@ public class Convert {
 	 * @return 替换后的字符
 	 */
 	public static String toDBC(String text, Set<Character> notConvertSet) {
+		if(StrUtil.isBlank(text)) {
+			return text;
+		}
 		char c[] = text.toCharArray();
 		for (int i = 0; i < c.length; i++) {
 			if (null != notConvertSet && notConvertSet.contains(c[i])) {